We can now purchase kingdoms.
- When a kingdom modal is open and the kingdom becomes npc the modal
  properly updates.
- When a user purchses a kingdom, the modal updates.

- Added a way to recalculate High Players next level neded XP.
  - This is in line with the new formula above level 1000.
  - The formula is: ((nextlevel - 1000) * 10) + 100 = next level XP
    requirment.
    - This means the XP needed past level 1000 will scale.

- Skill XP has changed: now its 5 xp per kill for trainable + bonuses
  - The amount needed per level is next level * 10.

Various other changes.

+++ app/Flare/Builders/Character/AttackDetails/DamageDetails/DamageSpellInformation.php
@@ -63,7 +63,7 @@ class DamageSpellInformation {
 
         $damage = $this->calculateClassSpellDamage($character, $damage, $voided);
 
-        return $damage + $damage * $bonus;
+        return (int) ($damage + $damage * $bonus);
     }
 
     /**
